refactor: add text in strings.xml and adjust spacing layout
