+++ app/src/main/java/com/app/desafiodourado/feature/home/ui/components/InfoComponent.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMaterial3Api::class)
 
-package com.app.desafiodourado.feature.home.ui
+package com.app.desafiodourado.feature.home.ui.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
